create add to library page
