+++ src/screens/library/LibraryPlaylistRow.js
@@ -6,31 +6,14 @@ import {
   TouchableHighlight,
 } from "react-native";
 import { COLORS } from "constants/theme";
-import { useNavigation } from "@react-navigation/native";
-import { LIBRARY_FLOW } from "constants/values";
-import { useSelector } from "react-redux";
-import { getAllSongByDocId } from "api/LibraryAPI";
 
-export default function LibraryPlaylistRow({ title, image, numOfSong, id }) {
-  const navigation = useNavigation();
-  const { lovedSongId, currLovedSong } = useSelector((state) => state.player);
-
-  const onPress = async () => {
-    const data =
-      lovedSongId === id ? currLovedSong : await getAllSongByDocId(id).songs;
-
-    navigation.navigate("PlaylistPage", {
-      id: id,
-      type: LIBRARY_FLOW,
-      props: {
-        image: image,
-        title: title,
-        numOfSong: numOfSong,
-        songs: data,
-      },
-    });
-  };
-
+export default function LibraryPlaylistRow({
+  onPress,
+  title,
+  image,
+  numOfSong,
+  id,
+}) {
   return (
     <TouchableHighlight
       underlayColor={COLORS.songRowClickColor}
